pyproject.toml
[project]
name = "PythonProject5"
version = "0.1.0"
description = "Project for testing API of http://dashboard.fast-api.senior-pomidorov.ru"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "allure-pytest>=2.15.2",
    "dotenv>=0.9.9",
    "faker>=38.2.0",
    "pytest>=9.0.2",
    "requests>=2.32.5",
]
